Move src/context -> src/common/context
